+++ pubspec.yaml
@@ -28,6 +28,8 @@ dependencies:
   # The following adds the Cupertino Icons font to your application.
   # Use with the CupertinoIcons class for iOS style icons.
   cupertino_icons: ^0.1.3
+  http: ^0.12.2
+
 
 dev_dependencies:
   flutter_test:
